requirements.txt
appnope==0.1.0
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
decorator==4.4.2
Django==3.1
django-elasticsearch-dsl==7.1.4
djangorestframework==3.11.1
elasticsearch==7.8.1
elasticsearch-dsl==7.2.1
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
